Revert "Added many more classes to help with UML class diagram.  Organized classes from each tier of the 3-tier architecture into separate folders."
This reverts commit cf2d94553761064cb6b3a47e4b9739f611c17b95.
