+++ src/trackit/ISQLHelper.java
@@ -1,4 +1,4 @@
-package trackit.DAL;
+package trackit;
 
 import java.util.*;
 import java.sql.*;
