+++ src/trackit/SQLConnector.java
@@ -1,4 +1,4 @@
-package trackit.DAL;
+package trackit;
 
 import java.sql.*;
 
@@ -8,7 +8,7 @@ import java.sql.*;
 public class SQLConnector {
 
     // <editor-fold defaultstate="collapsed" desc="Constants">
-    private final String DRIVER = "com.mysql.jdbc.Driver";
+    private String DRIVER = "com.mysql.jdbc.Driver";
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Private Fields">
     /**
@@ -90,7 +90,7 @@ public class SQLConnector {
      * @return True = the connection string is correct; False = the connection
      * string is incorrect.
      */
-    public boolean isValidConnection() {
+    public boolean validateConnection() {
         boolean isValid = false;
 
         return isValid;
