+++ src/trackit/SQLHelperSuppliers.java
@@ -1,29 +1,28 @@
-package trackit.DAL;
+package trackit;
 
 import java.sql.*;
 import java.util.*;
-import trackit.Item;
 
 /**
- * Handles all the CRUD operations for Items.
+ * Handles all the CRUD operations for Suppliers.
  */
-public class SQLHelperItems
-        extends SQLHelper<Item>
-        implements ISQLHelper<Item> {
+public class SQLHelperSuppliers
+        extends SQLHelper<Supplier>
+        implements ISQLHelper<Supplier> {
 
     /**
      *
      * @return @throws SQLException
      */
     @Override
-    public ArrayList<Item> selectAll()
+    public ArrayList<Supplier> selectAll()
             throws SQLException {
-        ArrayList<Item> results = new ArrayList<>();
+        ArrayList<Supplier> results = new ArrayList<>();
 
         HashMap<Integer, String> params = new HashMap<>();
-        ResultSet rs = super.execSproc("sp_Items_Select", params);
+        ResultSet rs = super.execSproc("sp_Suppliers_Select", params);
         while (rs.next()) {
-            Item anObj = new Item();
+            Supplier anObj = new Supplier();
             //Populate anObj properties from rs.
             results.add(anObj);
         }
@@ -32,31 +31,31 @@ public class SQLHelperItems
     }
 
     @Override
-    public ArrayList<Item> selectOne(Integer primaryKey)
+    public ArrayList<Supplier> selectOne(Integer primaryKey)
             throws SQLException {
         return new ArrayList<>();
     }
 
     @Override
-    public void insertAll(List<Item> aList)
+    public void insertAll(List<Supplier> aList)
             throws SQLException {
 
     }
 
     @Override
-    public void insert(Item anObject)
+    public void insert(Supplier anObject)
             throws SQLException {
 
     }
 
     @Override
-    public void updateAll(List<Item> aList)
+    public void updateAll(List<Supplier> aList)
             throws SQLException {
 
     }
 
     @Override
-    public void update(Item anObject)
+    public void update(Supplier anObject)
             throws SQLException {
 
     }
